pyproject.toml
[tool.poetry]
name = "data-dictionary"
version = "0.1.0"
description = ""
authors = ["Dan Lu, Nicholas Lee"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
mdutils = "==1.6.0"
numpy = "==1.24.3"
pandas = "==2.0.1"
python-frontmatter = "==1.0.0"
ipykernel = "^6.25.2"
jinja2 = "^3.1.2"
markdown = "^3.4.4"
tabulate = "^0.9.0"
synapseclient = "^3.0.0"
openpyxl = "^3.1.2"
linkml = "^1.6.1"


[tool.poetry.group.dev.dependencies]
pylint = "^2.17.5"
mypy = "^1.5.1"
autoflake = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
